% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_fetch.R
\name{shape_fetch}
\alias{shape_fetch}
\title{Get shapefiles from Natural Earth.}
\usage{
shape_fetch(ne.type = "physical", ne.filename = "ne_110m_land",
  ne.scale = "110m", dest = "./data/natural-earth")
}
\arguments{
\item{ne.type}{Set the type of map (see \url{http://www.naturalearthdata.com/downloads/}
for details). Supports: 'physical' and 'cultural' (default value is
'physical').}

\item{ne.filename}{Set the map layer (see \url{http://www.naturalearthdata.com/downloads/}
for details). Default value is 'ne_110m_land'.}

\item{ne.scale}{Set the map scale (see \url{http://www.naturalearthdata.com/downloads/}
for details). Supports: '10m' (1:10m), '50m' (1:50m), and '110m' (1:110m)
(default value is '110m'). Please note that '10m' file size is large.}

\item{dest}{Set the path of the destination folder the downloaded zip file
will be extracted into (default path is './data/natural-earth').}
}
\description{
Import shapefiles from \url{http://www.naturalearthdata.com}
and extract the contents in preperation for converting the shapefile to
GeoJSON using \code{geojson_convert}.
}

